Allow higher precision temp buffer in blend tests
Some blend tests were disabled because the results differ on
implementations that store the intermediate values in higher precision,
and only store the final result in the target buffer format.

This change enables the tests, and adds two additional legal results for
the tests. If none of the channels in the target buffer format exceed
8 bits, comparison against a 8888 buffer result is checked. If that
fails, or some channel is wider than 8 bits, comparison against a F64
buffer is done.

The 8888 and F64 results are first converted to the target buffer
format, and comparison threshold is the same as with the direct
comparison (dependent on the target buffer precision).

The higher precision tests are not done for formats where all channels
fit into 8, 16 or 32 bit formats.

Affects:

dEQP-VK.pipeline.blend.format.*

Components: Vulkan

VK-GL-CTS issue: 105

Change-Id: I26d99df151d9602bf9600b1694c3165990b93841