Keep buffer memory alive through BufferView creation
The device memory used to create a buffer was going out
of scope and being freed too early. A buffer view was
subsequently made using the buffer. This is invalid.

This change extends the lifetime of the buffer memory to
match the buffer.

Affects:

dEQP-VK.api.buffer_view.create.*

Components: Vulkan

VK-GL-CTS issue: 574

Change-Id: I6605da750d6494587781285447a157ef12018459