Support Android NDK r15
Most of the changes vs. r11 are for dealing with the the NDK's new
Unified Headers: there is now just one set of headers for all ABIs and
all API versions, but libraries are still per-ABI and per-API.

Building with NDK r15 requires CMake 3.2; otherwise, CMake 2.8 is
still supported.

Components: Framework
Change-Id: I4e64b4516c5ac9f59ac97adf331e01fa8be8eae0