Do not use degenerate frames in EGL tests
The dEQP-EGL.functional.buffer_age.* and
dEQP-EGL.functional.partial_update.* tests use randomly-generated frames
for rendering. The test allows the generation of frames with a width or
height of 0, but the tcu::PixelBufferAccess interface does not permit
0-sized subregions (see assertions inside getSubregion overloads in
tcuTextureUtil.cpp).

Skip the rendering of degenerate frames in the reference renderer, as
they are unsupported but would not affect the rendering if they were.

Change-Id: I3977c3c293715db633dde74132e56b5d9303ab28