Clear all attachment layers for the multiview tests
We check them all in the end, even though a clear only touches the
layers which are rendered to as views in the renderpass.

As we do not clear the entire image, and we still check all the
unused layers for 0, this fails randomly on drivers that don't
zero-initialize memory.

This adds the clear instead of removing the checks, so we keep some
superficial check that the unused layers aren't modified.

Affects:

dEQP-VK.multiview.*

Components: Vulkan

VK-GL-CTS issue: 630

Change-Id: I48f46708c983f9f273affe223ddc06c7a95cec88