Fix flakiness in mailbox.display_timing
The "vktWsiDisplayTimingTests.cpp" file was originally written just for
a present mode of VK_PRESENT_MODE_FIFO_KHR.  It has a case at image 80,
where it changes the desiredPresentTime to be 1 second earlier, and then
checks that later.  Such a check doesn't work with
VK_PRESENT_MODE_MAILBOX_KHR, which can throw off the timing of images
(because some images may not be presented).  This has resulted in the
test having random failures.  The "image 80 test" wasn't intended to
occur with MAILBOX, and so it's now being removed.

VK-GL-CTS issue: 622
Components: Vulkan
Affects: dEQP-VK.wsi.*.display_timing.mailbox.display_timing
Bug: 64486630
Test: Build cts and ran scripts/check_build_sanity.py

Change-Id: I9fbd2254b4e6fc1c7bff077c8c0a5229a6d74800