Use hidden visibility with GCC and Clang
The final result of the build are standalone executables, with all the
code statically linked.

Change the visibility to hidden, to avoid the symbol leakage and allow
the compiler to discard some code.

As a result the binary size is drastically decreased.

Total executable size on my setup drops by ~20%
 158M    before
 128M    after

Note: The binaries still export multiple unneeded symbols. Resolving
those is left as an exercise for later.

Test: full rebuild, running resulting binaries.

Change-Id: Ife10bbd123d59f4480349794785f13978516df30
Signed-off-by: Emil Velikov <emil.velikov@collabora.com>