Restore fuzzyCompare() sample rate
Commit 57c002a31612855e744c33ac698590da0e3d02d3 changed how the stride
was calculated for sampling values in order to avoid double-counting
values. However, it also reduced the number of values sampled and
that's resulting in both GLES and Vulkan failures.

This patch restores the original stride calculation, but changes the
parameters to rnd.getInt() to ensure that 0 is never returned from it.

Components: Framework

VK-GL-CTS issue: 658

Change-Id: I4c7e6d39017bbfc38bcd155037d3b4d40ce46ca7