Handle edge case in GL FP64 tests
mod(a,b) is defined as: a - b * floor(a/b)
mod(a,a) should be 0 however OpenGL allows for some error in
division, so a/a could actually end up being 1.0 - 1ULP. In
such a case, floor(a/a) would end up as 0, with mod(a,a)
becoming a. This patchset detects this case and adjusts the
expected result accordingly.

Components: OpenGL

VK-GL-CTS issue: 641

Affects:
KHR-GL40.gpu_shader_fp64.builtin.*

Change-Id: I9535d337b2132519c985f8ac8eb18bf938a89ea7