Lower correlation threshold in flush-finish tests again
This test expects predictability from glFlush/glFinish timing.
In particular,  the code from FlushFinishCase::analyzeResults()
checks that the result falls within specific timing thresholds.

Notice that in the GLES3 specification, there is no requirement for
timing, just a guarantee that "commands [...] complete in finite time".
This is a problem because it makes the test very flaky, for example
if the CPU/GPU clocks are not ramping up fast enough.

BUG=b:67331095
TEST=Ran dEQP-GLES3.functional.flush_finish.finish on Elm for over 3+ hours
     Observed 1.5% to 2% flakiness without this patch. Observed no flakiness
     with this patch.