Ignore the patch number when comparing apiVersions
Modify the getSpirvVersionForAsm() and getSpirvVersionForGlsl()
functions to ignore the patch version number when checking
for Vulkan 1.0 vs Vulkan 1.1. It's often the case that the loader
or implementation will have a non-zero patch number to indicate
what version of the spec and vulkan header they built with.

Affects: startup

Components: Vulkan

VK-GL-CTS issue: 767

Change-Id: I947cb43db5e1f93e5b0b9cb93d205863b8e3c4b1