Fix for getPhysicalDeviceFeatures2 loader issue
For all instance methods taking VkPhysicalDevice
as a first parameter that were moved to core in
1.1 there is now a dynamic dispatch based on
physical device api version. If version is 1.1 or
above a core method is used. If version is lower
than that a counterpart extension method is used.

Additionally a pre-instance is created to enumerate
all physical devices and select lowest version to
be able to enable proper extensions set for final
instance.

Components: Vulkan

VK-GL-CTS issue: 854

Affects: All tests using getPhysicalDeviceFeatures2

Change-Id: Ia68a5cac23ae35566fcf99bc56c93c9cb77f33e0