Set swap interval to 0
By default, when swapping front and back frame buffers, this is in sync
with v-blank, which prevents tearing.

So this means the tests does not run at is maximum allowed speed, but it
is restricted somewhat by the attached screen.

When running in a window and X realizes the is no monitor attached, the
server's Present code will vsync at 1 FPS, which artificially slow down
the tests a ton.

To avoid this issue, and also because we want to run things as fast as
possible, set the swap interval to 0, which basically deactivate the
sync.

Test: cmake .. -DDEQP_TARGET=x11_egl

Change-Id: I04bcb814fb289d58f8788e90f5118db0b7175777