glSpirvTests: fix transform feedback query
The test removes the names for several variables of the shader, and
then it expect a length of 1 for all of them when doing a state
query.

But on the case of TRANSFORM_FEEDBACK_VARYING_MAX_LENGTH it is not
activating any varying.

This commits adds an output definition with xfb decorations so the
variable would become active without the need to call
glTransformFeedbackVaryings. From ARB_gl_spirv spec:

    "If the shader used to record output variables for transform
    feedback varyings uses the xfb_buffer, xfb_offset, or xfb_stride
    layout qualifiers, or its SPIR-V entry point declares the *Xfb*
    Execution Mode, the values specified by TransformFeedbackVaryings
    are ignored, and the set of variables captured for transform
    feedback is instead derived from the specified layout qualifiers
    or SPIR-V decorations: outputs specifying both an *XfbBuffer* and
    an *Offset* are captured, while outputs not specifying both of
    these are not captured. Values are captured each time the shader
    writes to such a decorated object."

As we are here, we also simplify how name reflection is
stripped. Instead of stripping each individual OpName and OpMemberName
that we know the shader have, we just remove any OpName and
OpMemberName line.

Modules: OpenGL

VK-GL-CTS issue: 886

Affects:
KHR-GL45.gl_spirv.spirv_modules_state_queries_test
KHR-GL46.gl_spirv.spirv_modules_state_queries_test

Change-Id: I124c90c61f8bee1e873867cf0d2b2e689679d70a