Check for depth/stencil buffer in default FBO
The affected tests invoke gl.getNamedFramebufferAttachmentParameteriv()
with GL_FRAMEBUFFER_ATTACHMENT_DEPTH_SIZE to get the depth bits used in
the default FBO. And something similar is done for the stencil bits.

But wrongly it assumes that if the FBO does not have an associated depth
or stencil buffer, it would return 0. Rather, this function returns an
error if it is invoked in a FBO without a depth/stencil buffer attached.

So better check first if the default FBO has a depth/stencil buffer
attached, and if so then invoke the above functions to get the
depth/stencil bits.

Components: OpenGL

VK-GL-CTS issue: 969

Affects:
KHR-*.internalformat.renderbuffer.*

Change-Id: I74d28d5d089790fed4dd2728cb5143e7c03d7442