SparseTextureAlloc check for GL_NV_deep_texture3D
When running the
SparseTextureAllocationTestCase::verifyTexStorageInvalidValueErrors test
we need to check if the GL_NV_deep_texture3D extension is supported. If
so we can't compare the depth of the texture against the
GL_MAX_3D_TEXTURE_SIZE attribute that is intended for symmetric texture
limits. Instead we need to check against GL_MAX_DEEP_3D_TEXTURE_DEPTH_NV
provided the texture width and height are within the acceptable ranges
for a deep texture.

Components: OpenGL

VK-GL-CTS issue: 959

Change-Id: I0e404d6851724b0624d0802603a12aea3af220e5
Reviewed-by: ewerness