GLES2: fix typo in es2fShaderStructTests
I happened to notice this typo while editing the file. It could cause
some tests to be skipped when they should not be: it skips a case if
(1) the device has no vertex texture image units and
(2) the case uses any flags at all and
(3) the case is a vertex case.

There could be conformance impact from this change.

Affects:
  dEQP-GLES2.functional.shaders.struct.uniform.array_member_dynamic_index_vertex
  dEQP-GLES2.functional.shaders.struct.uniform.struct_array_dynamic_index_vertex
  dEQP-GLES2.functional.shaders.struct.uniform.nested_struct_array_dynamic_index_vertex
  dEQP-GLES2.functional.shaders.struct.uniform.dynamic_loop_struct_array_vertex
  dEQP-GLES2.functional.shaders.struct.uniform.dynamic_loop_nested_struct_array_vertex
Components: AOSP

Change-Id: I5110d6c240938bc58664e7dd5ee2b43e26ceea6e