Removed incompatible render pass test
Incompatible render pass test was based on a validation layer test
VkLayerTest.RenderPassIncompatible which turned out to be invalid.
Render pass creation in that test was also deemed invalid by the
validation layer, which caused an uninitialized render pass object
to be used. This caused a crash on some devices.

Components: Vulkan

VK-GL-CTS issue: 841

Change-Id: I9029d5bb7fa9c6be1e359b73d8bfeac3716bfc88