Added optimization avoiding shader disassembly
This change avoids disassembling shaders if the
--deqp-log-shader-sources=disable option is used.

The default functionality disassembles everything and discards the
result if the shader source logging is disabled. This change skips the
disassembly too.

The performance gain for this is approximately 10% of runtime.

The timings are based on assumption of debug build and having the
shader cache with zero cache misses.

Affects:

none

Components: Framework

Vk-GL-CTS isue: 899

Change-Id: I36eda000a94400d8f50a565c7ef30f4170d51c83