Test using a pipeline whose render pass has been destroyed
This test has the same functionality as
VkPositiveLayerTest.DestroyPipelineRenderPass.

The test creates two compatible render passes. The first one is used
to create the graphics pipeline and the second one is used for
everything else. The first render pass is destroyed while the second
render pass is active and the graphics pipeline is bound to a
command buffer.

The test passes as long as no crash happens.

New tests:

dEQP-VK.api.pipeline.renderpass.destroy_pipeline_renderpass

Components: Vulkan

VK-GL-CTS issue: 833

Change-Id: I7eff6595da6dcfacfe8c49b14c2f9a36d5436088