DO NOT MERGE: Fix dEQP-VK.api.copy_and_blit.image_to_image_depth value range.
Per spec:

"Because depth or stencil aspect buffer to image copies may require format conversions on some
implementations, they are not supported on queues that do not support graphics. When copying to
a depth aspect, the data in buffer memory must be in the the range [0,1] or undefined results occur."

Initializing it to integer values falls afoul of this rule
quickly. As the image is 256 by 256, x/255.0 should always fall in the range [0,1].

Not entirely sure this would initialize combined depth stencil correctly,
but this version of CTS does not have a test using those formats with
this code.

This has been fixed upstream for a while with

"9909b9f3 Fix combined depth/stencil clear colors."

but this is not a straight backport as these tests saw some significant rework between
this CTS version and the fix.

Bug: b/73449389

Change-Id: Iaf0ce14ed4338d437f1b20033be5d0f8c0cbd699