Restrict alpha range for rgba5551 surfaces
The tests were blending against an rgba5551 surface and expecting alpha
to be quantized to 1 bit between every drawcall. The spec does not
forbid implementations to use higher precision internal representations
where this quantisation may not occur to the same extent. This change
restricts fragment outputs to be either 0 or 1, so the issue is avoided

Components: OpenGL

VK-GL-CTS issue: 988

Affects:
dEQP-EGL.functional.render.*rgba5551*

Change-Id: Id8fa5068931124d79ae7fa1b9b8a41845c73e507