RGB10_A2UI renderbuffer case requires integer data
Most of the cases in KHR-GL40.internalformat.renderbuffer.* test are
float-based, and thus the shaders as well as the functions used to read
the pixels are based on floats.

But for this specific format, we need to use integer data, both when the
shader writes out the colors as well as when we read it back with
readPixels(). And later we need to convert the results to floats in
order to compare with the expected results (which are float-based
values).

Components: OpenGL

VK-GL-CTS issue: 1001

Affects:
KHR-GLES*.core.internalformat.renderbuffer.*
KHR-GL*.internalformat.renderbuffer.*

Change-Id: I44c051817672614fb1ed6ed1927f917afcaadc96