Test waiting for fences without waitAll
Two command buffers and two fences are created. Both command buffers
are submitted to a queue, and vkWaitForFences with waitAll set to
VK_FALSE is used to wait for either one of the fences to be signaled.

New tests:

dEQP-VK.synchronization.basic.fence.multi_waitall_false

Components: Vulkan

VK-GL-CTS issue: 1037

Change-Id: I78762b2561e8d28a9a4278a054cbe169af10366a