VulkanPlatform for OSX
The change modularizes the OSX platform framework by splitting the GL
platform implementation out (into tcu::OSXGLPlatform) and adding a
basic Vulkan platform implementation (in tcu::osx::VulkanPlatform).
It is currently missing the WSI integration, and the memory limits may
need to be revised in follow-ups. This change allows the CTS to be run
against a Vulkan portability library on MacOS.

Components: Framework

Change-Id: I67eb9af5b97de701b4188aedcaca7485276d9001