Add CONTRIBUTING.md for Github
Developers used to Github workflow may be tempted to make a PR change,
and it's important to point them to the guidelines as early as possible
in order to avoid the bounced response.
The message they'll see will look like this:
https://help.github.com/articles/setting-guidelines-for-repository-contributors/

Affects: users making Pull Requests on Github.

Components: none

Change-Id: I5921bb4ae4cb8a28f889f3f20c134ca82784c3f2