Add tests for checking YCbCr format enums in Vulkan 1.1
Check that YCbCr format enums added in Vulkan 1.1 are
recognized by implementation, even when samplerYcbcrConversion
feature is not supported.

This change splits 'checkYcrbrConversionSupport' into:
- API support (if VkFormats are recognized by the implementation)
- the actual conversion support.

YCbCr feature check tests are run always when API is supported,
even if YCbCr conversion it not. However, support for format properties
and image format properties is still not required, if YCBCr conversion feature is
not supported.

Component: Vulkan

VK-GL-CTS Issue: 1059

Affects:
	dEQP-VK.api.info.format_properties.*
	dEQP-VK.api.info.*image_format_properties*

Change-Id: I84280eaa481a566e05ce028b9e4ec407f2d27599